# services/hlims/generated/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package database

import (
	"database/sql"
)

type Link struct {
	ID       string `json:"id"`
	Short    string `json:"short"`
	Long     string `json:"long"`
	Created  int64  `json:"created"`
	Lastedit int64  `json:"lastedit"`
	Owner    string `json:"owner"`
}

type Stat struct {
	ID      string        `json:"id"`
	Created int64         `json:"created"`
	Clicks  sql.NullInt64 `json:"clicks"`
}